themes, layouts, content: Fixup style
The old version of the theme wasn't showing a lot of common markdown
patterns like italics and backticks and so on... Updating the theme to
the newest version broke everything, so this first fixes a bunch of
things (including some hacks) and then fixes up the content to work
properly.

Yuck!